Move traverse into servicecatalog/client
