--- pkg/traverse/instance.go
@@ -1,96 +1,0 @@
-package traverse
-
-import (
-	"fmt"
-
-	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
-	"github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset"
-	"k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-// InstanceToServiceClassAndPlan retrieves the parent class and plan for an instance.
-func InstanceToServiceClassAndPlan(
-	cl *clientset.Clientset,
-	instance *v1beta1.ServiceInstance,
-) (*v1beta1.ClusterServiceClass, *v1beta1.ClusterServicePlan, error) {
-	classID := instance.Spec.ClusterServiceClassRef.Name
-	classCh := make(chan *v1beta1.ClusterServiceClass)
-	classErrCh := make(chan error)
-	go func() {
-		class, err := cl.ServicecatalogV1beta1().ClusterServiceClasses().Get(classID, v1.GetOptions{})
-		if err != nil {
-			classErrCh <- err
-			return
-		}
-		classCh <- class
-	}()
-
-	planID := instance.Spec.ClusterServicePlanRef.Name
-	planCh := make(chan *v1beta1.ClusterServicePlan)
-	planErrCh := make(chan error)
-	go func() {
-		plan, err := cl.ServicecatalogV1beta1().ClusterServicePlans().Get(planID, v1.GetOptions{})
-		if err != nil {
-			planErrCh <- err
-			return
-		}
-		planCh <- plan
-	}()
-
-	var class *v1beta1.ClusterServiceClass
-	var plan *v1beta1.ClusterServicePlan
-	for {
-		select {
-		case cl := <-classCh:
-			class = cl
-			if class != nil && plan != nil {
-				return class, plan, nil
-			}
-		case err := <-classErrCh:
-			return nil, nil, err
-		case pl := <-planCh:
-			plan = pl
-			if class != nil && plan != nil {
-				return class, plan, nil
-			}
-		case err := <-planErrCh:
-			return nil, nil, err
-
-		}
-	}
-}
-
-// InstanceParentHierarchy retrieves all ancestor resources of an instance.
-func InstanceParentHierarchy(cl *clientset.Clientset, instance *v1beta1.ServiceInstance,
-) (*v1beta1.ClusterServiceClass, *v1beta1.ClusterServicePlan, *v1beta1.ClusterServiceBroker, error) {
-	class, plan, err := InstanceToServiceClassAndPlan(cl, instance)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	broker, err := ClassToBroker(cl, class)
-	if err != nil {
-		return nil, nil, nil, err
-	}
-
-	return class, plan, broker, nil
-}
-
-// InstanceToBindings retrieves all child bindings for an instance.
-func InstanceToBindings(cl *clientset.Clientset, instance *v1beta1.ServiceInstance,
-) ([]v1beta1.ServiceBinding, error) {
-	// Not using a filtered list operation because it's not supported yet.
-	results, err := cl.ServicecatalogV1beta1().ServiceBindings(instance.Namespace).List(v1.ListOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("unable to search bindings (%s)", err)
-	}
-
-	var bindings []v1beta1.ServiceBinding
-	for _, binding := range results.Items {
-		if binding.Spec.ServiceInstanceRef.Name == instance.Name {
-			bindings = append(bindings, binding)
-		}
-	}
-
-	return bindings, nil
-}
